#65 added base runner. making more generic
