+++ src/conversion/ComponentExtractionStrategy.js
@@ -1,11 +1,11 @@
-import ExtendableExtractionStrategy from './ExtendableExtractionStrategy';
+import ExtendableComponentExtractionStrategy from './ExtendableComponentExtractionStrategy';
 
 export default class ComponentExtractionStrategy {
 	constructor(data) {
 		this.data = data;
 		this.keyword = 'component';
 		this.containerName = 'elements';
-		this.parentStrategy = new ExtendableExtractionStrategy(data, {
+		this.parentStrategy = new ExtendableComponentExtractionStrategy(data, {
 			keyword: 'component',
 			containerName: 'elements',
 		});
