+++ src/conversion/SubMapExtractionStrategy.js
@@ -1,9 +1,9 @@
-import ExtendableExtractionStrategy from './ExtendableExtractionStrategy';
+import ExtendableComponentExtractionStrategy from './ExtendableComponentExtractionStrategy';
 
 export default class SubMapExtractionStrategy {
 	constructor(data) {
 		this.data = data;
-		this.parentStrategy = new ExtendableExtractionStrategy(data, {
+		this.parentStrategy = new ExtendableComponentExtractionStrategy(data, {
 			keyword: 'submap',
 			containerName: 'submaps',
 		});
